Show the different versions of the manual pages
Note: While I tried very hard to find an efficient way to regenerate the
AsciiDoc sources on-the-fly only when they change (so they can be
diff'ed), I failed in the endeavor: There are too many includes to make
that practical (read: generated sources can change in case a number of
files change, not just `Documentation/git-<command>.txt`).

So now we have this big, ugly `_generated-asciidoc/` directory that is
full of files whose names are their SHA-1s. If it ever becomes a problem
(for example because there might be too many files in that directory),
we will have to revisit this strategy.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>